refactor(core): remove utils/storage.ts; move store factory + registry into store/
The `Storage` class is gone; `utils/storage.ts` was left holding the store factory
and the instance registry, which no longer fit that file/name. Split them into the
store module: `store/createStore.ts` (createVueFlowStore factory) and
`store/registry.ts` (createFlowStore/getFlowStore/generateFlowId — the useVueFlow(id)
escape hatch). `useVueFlow` imports the registry directly. No behaviour change.
